Make StringValue's memory layout same with Slice
In our storage engine's code, we cast StringValue to Slice. Because
their memory layout is not differenct, it may cause BE process crash.

We make their memory layout same in this patch to resolve this problem
temporary. We should improve it some day.

+++ be/src/exec/text_converter.cpp
@@ -37,7 +37,7 @@ void TextConverter::unescape_string(StringValue* value, MemPool* pool) {
     value->ptr = new_data;
 }
 
-void TextConverter::unescape_string(const char* src, char* dest, int* len) {
+void TextConverter::unescape_string(const char* src, char* dest, size_t* len) {
     char* dest_ptr = dest;
     const char* end = src + *len;
     bool escape_next_char = false;
